+++ browser/components/navbar.jsx
