+++ browser/components/player.jsx
