+++ browser/index.jsx
@@ -1,12 +1,13 @@
-/* global document */
-
+/* globals document */
+import 'bootstrap/dist/css/bootstrap.css';
+import 'bootstrap/dist/css/bootstrap-theme.css';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { render } from 'react-dom';
 import { Router, Route } from 'react-router';
 import { Provider } from 'react-redux';
-import store, { history } from './redux/store';
-
-const App = () => (<span>Hello World!</span>);
+import store, { history } from './redux/store'; //Change
+import App from './components/app';
+import './index.css';
 
 const router = (
   <Provider store={store}>
@@ -16,5 +17,5 @@ const router = (
   </Provider>
 );
 
-ReactDOM.render(router, document.getElementById('root'));
+render(router, document.getElementById('root'));
