+++ browser/redux/reducers/application.jsx
