+++ browser/redux/reducers/chat.jsx
