+++ browser/redux/reducers/controls.jsx
