+++ browser/redux/reducers/index.jsx
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
+import { pip } from './pip-player';
+import { controls } from './controls';
+import { chat } from './chat';
+import { application } from './application';
 
-const rootReducer = combineReducers({
-  routing: routerReducer,
-});
+
+const rootReducer = combineReducers({ application, chat, pip, controls, routing: routerReducer });
 
 export default rootReducer;
-
