+++ browser/redux/reducers/pip-player.jsx
