+++ package.json
@@ -24,16 +24,20 @@
     "cookie-parser": "^1.4.3",
     "express": "^4.14.0",
     "express-session": "^1.14.1",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "file-loader": "^0.10.0",
     "helmet": "^3.0.0",
     "morgan": "^1.7.0",
+    "muicss": "^0.9.9-rc2",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "pg": "^6.1.0",
     "pg-native": "^1.10.0",
     "react": "^15.4.1",
-    "react-bootstrap": "^0.30.3",
+    "react-bootstrap": "^0.30.7",
     "react-dom": "^15.4.1",
     "react-draggable": "^2.2.2",
+    "react-fa": "^4.1.2",
     "react-fontawesome": "^1.2.0",
     "react-redux": "^4.4.6",
     "react-router": "^3.0.0",
@@ -43,6 +47,7 @@
     "redux": "^3.6.0",
     "redux-thunk": "^2.1.0",
     "sequelize": "^3.24.1",
+    "url-loader": "^0.5.7",
     "webpack": "^1.13.3"
   },
   "devDependencies": {
