+++ public/index.html
@@ -5,13 +5,16 @@
     <meta charset="utf-8">
     <title>Twitchception</title>
     <meta name="viewport" content="initial-scale=1, maximum-scale=1">
-    <link rel='stylesheet' href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css" />
+    <link rel='stylesheet' href="/public/dist/bundle.css" />
+    <link rel='stylesheet' href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" />
+    <link href="//cdn.muicss.com/mui-0.9.0/css/mui.min.css" rel="stylesheet" type="text/css">
+    <script src="//cdn.muicss.com/mui-0.9.0/js/mui.min.js"></script>
   </head>
 
   <body>
     <div id="root"></div>
     <!-- Scripts -->
-    <script src="http://player.twitch.tv/js/embed/v1.js"></script>
+    <script src="https://player.twitch.tv/js/embed/v1.js"></script>
     <script src="/public/dist/bundle.js"></script>
   </body>
