--- src/index.js
@@ -1,22 +1,0 @@
-import 'bootstrap/dist/css/bootstrap.css';
-import 'bootstrap/dist/css/bootstrap-theme.css';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { render } from 'react-dom';
-import { Router, Route, IndexRoute, browserHistory } from 'react-router';
-import { Provider } from 'react-redux';
-import store, { history } from './redux/store'; //Change
-import App from './components/app';
-import './index.css';
-
-const router = (
-	<Provider store={store}>
-		<Router history={history}>
-			<Route path="/" component={App}>
-			</Route>
-		</Router>
-	</Provider>
-)
-
-render(router,document.getElementById('root'));
-
